dom/indexedDB/test/test_abort_on_reload.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Test</title>

  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript">
    let openedWindow
    let reloads = 0

    function openWindow() {
      openedWindow = window.open("abort_on_reload.html");
    }

    function messageListener() {
      ok(true, "reload recorded");

      if (++reloads == 20) {
        openedWindow.close();
        SimpleTest.finish();
      }
    }

    function runTest() {
      SimpleTest.waitForExplicitFinish();

      window.addEventListener("message", messageListener);

      openWindow();
    }
  </script>

</head>

<body onload="runTest();">
</body>

</html>
